fix(comp): guard LENS_WINDOWS_PER_DAY to integers
A fractional value (e.g. 1.5) passed the previous z.coerce.number().min(1)
check and could inflate a window's accrued entitlement past the daily budget.

+++ src/infrastructure/config/env.ts
@@ -41,7 +41,7 @@ export const envSchema = z.object({
   LENS_MONTHLY_BUDGET: z.coerce.number().min(0).default(250),
   // The day's comp budget accrues one share per window rather than being fully
   // available at midnight. 4 windows = 6h, so a fresh find waits 6h at worst.
-  LENS_WINDOWS_PER_DAY: z.coerce.number().min(1).default(4),
+  LENS_WINDOWS_PER_DAY: z.coerce.number().int().min(1).default(4),
   // Candidates submitted to the comparative ranker per window. Absolute triage
   // scoring saturates (967 of 1651 qualified listings scored exactly 9 over the
   // two weeks to 2 Aug 2026), so the ranker, not the score, picks the winners.
